Separate interface from examples.
