+++ src/models/ability_examples.py
@@ -1,28 +1,7 @@
-"""Character abilities."""
-import abc
-
+from models.abilities_base import Ability
 from models.character_base import Character
 from models.conditions import FullHealth
 from models.states import Attribute
-
-
-class Ability(metaclass=abc.ABCMeta):
-
-    @abc.abstractmethod
-    def _use(self, user: Character, target: Character) -> None:
-        """Internal implementation of use method, must be overridden."""
-
-    @abc.abstractmethod
-    def can_use(self, user: Character, target: Character) -> bool:
-        """Whether the ability can be used."""
-
-    def use(self, user: Character, target: Character) -> None:
-        assert self.can_use(user, target)
-        self._use(user, target)
-
-    @abc.abstractmethod
-    def describe_use(self, user: Character, target: Character) -> str:
-        """Description of use for player."""
 
 
 class Repair(Ability):
